fix integer division problem in alarm

+++ src-cpp/session.cpp
@@ -11,8 +11,8 @@ Session::Session(int _minutes_per_sprint, int _target_wordcount,
 
 // parameters for mapping slider to seconds
 // TODO make constant
-severity_slider_max_value = 100;
-max_seconds_grace_period = 30;
+severity_slider_max_value = 100.0;
+max_seconds_grace_period = 30.0;
 
     text = _text;
 
@@ -45,11 +45,11 @@ void Session::update_session_status(QString current_text_string) {
 void Session::calc_minutes_remaining() {
 
     seconds_remaining = finishtime - time(NULL);
-    minutes_remaining = round(seconds_remaining / 60);
+    minutes_remaining = round(seconds_remaining / 60.0);
     minutes_remaining = std::max(minutes_remaining, 0);
 
     if (minutes_per_sprint > 0) {
-        perc_time_remaining = int(minutes_remaining / minutes_per_sprint * 100);
+        perc_time_remaining = int((float)minutes_remaining / (float)minutes_per_sprint * 100);
     } else {
         perc_time_remaining = 0;
     }
@@ -61,7 +61,7 @@ void Session::calc_word_stats() {
     words_remaining = target_wordcount - text.addedWordcount;
 
     if (target_wordcount > 0) {
-        perc_wc_achieved = int(text.addedWordcount / target_wordcount * 100);
+        perc_wc_achieved = int((float)text.addedWordcount / (float)target_wordcount * 100);
     }
     else {
         perc_wc_achieved = 0;
@@ -71,5 +71,5 @@ void Session::calc_word_stats() {
 int Session::calc_grace_period(int severity){
 //        Map from slider value to seconds
 
-    return int(severity / severity_slider_max_value * max_seconds_grace_period);
+    return int((float)severity / severity_slider_max_value * max_seconds_grace_period);
 }
